# apps/api/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/products": {
      "get": {
        "operationId": "ProductsController_findAll",
        "parameters": [
          {
            "name": "category",
            "required": false,
            "in": "query",
            "description": "Filtrar por slug de categoría",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Product"
                  }
                }
              }
            }
          }
        },
        "tags": [
          "products"
        ]
      },
      "post": {
        "operationId": "ProductsController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateProductDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "products"
        ]
      }
    },
    "/api/products/{id}": {
      "get": {
        "operationId": "ProductsController_findOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Product"
                }
              }
            }
          }
        },
        "tags": [
          "products"
        ]
      },
      "patch": {
        "operationId": "ProductsController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateProductDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "products"
        ]
      },
      "delete": {
        "operationId": "ProductsController_remove",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "products"
        ]
      }
    },
...
    "/api/categories": {
      "get": {
        "operationId": "CategoriesController_findAll",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Category"
                  }
                }
              }
            }
          }
        },
        "tags": [
          "categories"
        ]
      },
      "post": {
        "operationId": "CategoriesController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateCategoryDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Category"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "categories"
        ]
      }
    },
    "/api/categories/{slug}": {
      "get": {
        "operationId": "CategoriesController_findOne",
        "parameters": [
          {
            "name": "slug",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Category"
                }
              }
            }
          }
        },
        "tags": [
          "categories"
        ]
      }
    },
    "/api/categories/{id}": {
      "patch": {
        "operationId": "CategoriesController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateCategoryDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Category"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "categories"
        ]
      },
      "delete": {
        "operationId": "CategoriesController_remove",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "categories"
        ]
      }
    },
    "/api/product-variants": {
      "get": {
        "operationId": "ProductVariantsController_findAll",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "product-variants"
        ]
      },
      "post": {
        "operationId": "ProductVariantsController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateProductVariantDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "product-variants"
        ]
      }
    },
    "/api/product-variants/{id}": {
      "get": {
        "operationId": "ProductVariantsController_findOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "product-variants"
        ]
      },
      "patch": {
        "operationId": "ProductVariantsController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateProductVariantDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "product-variants"
        ]
      },
      "delete": {
        "operationId": "ProductVariantsController_remove",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "product-variants"
        ]
      }
    },
    "/api/attribute-types": {
      "get": {
        "operationId": "AttributeTypesController_findAll",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "attribute-types"
        ]
      },
      "post": {
        "operationId": "AttributeTypesController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateAttributeTypeDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "attribute-types"
        ]
      }
    },
    "/api/attribute-types/{id}": {
      "get": {
        "operationId": "AttributeTypesController_findOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "attribute-types"
        ]
      },
      "patch": {
        "operationId": "AttributeTypesController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateAttributeTypeDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "attribute-types"
        ]
      },
      "delete": {
        "operationId": "AttributeTypesController_remove",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "attribute-types"
        ]
      }
    }
  },
  "info": {
    "title": "API",
    "description": "",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
...
      "ProductCategory": {
        "type": "object",
        "properties": {
          "id": {
            "type": "number",
            "example": 1
          },
          "name": {
            "type": "string",
            "example": "Remeras"
          },
          "slug": {
            "type": "string",
            "example": "remeras"
          }
        },
        "required": [
          "id",
          "name",
          "slug"
        ]
      },
      "AttributeType": {
        "type": "object",
        "properties": {
          "id": {
            "type": "number",
            "example": 1
          },
          "name": {
            "type": "string",
            "example": "Talle"
          },
          "slug": {
            "type": "string",
            "example": "talle"
          }
        },
        "required": [
          "id",
          "name",
          "slug"
        ]
      },
      "AttributeValue": {
        "type": "object",
        "properties": {
          "id": {
            "type": "number",
            "example": 1
          },
          "value": {
            "type": "string",
            "example": "M"
          },
          "attribute_types": {
            "$ref": "#/components/schemas/AttributeType"
          }
        },
        "required": [
          "id",
          "value",
          "attribute_types"
        ]
      },
      "VariantAttribute": {
        "type": "object",
        "properties": {
          "attribute_values": {
            "$ref": "#/components/schemas/AttributeValue"
          }
        },
        "required": [
          "attribute_values"
        ]
      },
      "ProductVariant": {
        "type": "object",
        "properties": {
          "id": {
            "type": "number",
            "example": 1
          },
          "sku": {
            "type": "object",
            "example": "SKU-001"
          },
          "price_modifier": {
            "type": "object",
            "example": 0
          },
          "stock": {
            "type": "number",
            "example": 10
          },
          "variant_attributes": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/VariantAttribute"
            }
          }
        },
        "required": [
          "id",
          "sku",
          "price_modifier",
          "stock",
          "variant_attributes"
        ]
      },
      "Product": {
        "type": "object",
        "properties": {
          "id": {
            "type": "number",
            "example": 1
          },
          "name": {
            "type": "string",
            "example": "Remera Técnica Elite"
          },
          "price": {
            "type": "number",
            "example": 25.99
          },
          "image": {
            "type": "string",
            "example": "https://example.com/image.png",
            "nullable": true
          },
          "description": {
            "type": "string",
            "example": "Descripción del producto",
            "nullable": true
          },
          "category_id": {
            "type": "number",
            "example": 1
          },
          "created_at": {
            "type": "string",
            "example": "2026-04-10T00:00:00.000Z",
            "nullable": true
          },
          "categories": {
            "nullable": true,
            "allOf": [
              {
                "$ref": "#/components/schemas/ProductCategory"
              }
            ]
          },
          "product_variants": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ProductVariant"
            }
          }
        },
        "required": [
          "id",
          "name",
          "price",
          "image",
          "description",
          "category_id",
          "created_at",
          "categories",
          "product_variants"
        ]
      },
      "CreateProductDto": {
        "type": "object",
        "properties": {}
      },
      "UpdateProductDto": {
        "type": "object",
        "properties": {}
      },
...
      "Category": {
        "type": "object",
        "properties": {
          "id": {
            "type": "number",
            "example": 1
          },
          "name": {
            "type": "string",
            "example": "Remeras"
          },
          "slug": {
            "type": "string",
            "example": "remeras"
          },
          "description": {
            "type": "string",
            "example": "Remeras y camisetas",
            "nullable": true
          },
          "created_at": {
            "type": "string",
            "example": "2026-04-10T00:00:00.000Z",
            "nullable": true
          }
        },
        "required": [
          "id",
          "name",
          "slug",
          "description",
          "created_at"
        ]
      },
      "CreateCategoryDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "example": "Electronics"
          },
          "slug": {
            "type": "string",
            "example": "electronics"
          },
          "description": {
            "type": "string",
            "example": "Electronic gadgets and devices"
          }
        },
        "required": [
          "name",
          "slug"
        ]
      },
      "UpdateCategoryDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "example": "Electronics"
          },
          "slug": {
            "type": "string",
            "example": "electronics"
          },
          "description": {
            "type": "string",
            "example": "Electronic gadgets and devices"
          }
        }
      },
      "CreateProductVariantDto": {
        "type": "object",
        "properties": {
          "product_id": {
            "type": "number",
            "example": 1
          },
          "sku": {
            "type": "string",
            "example": "SKU-001"
          },
          "price_modifier": {
            "type": "number",
            "example": 0
          },
          "stock": {
            "type": "number",
            "example": 10
          },
          "is_active": {
            "type": "boolean",
            "example": true
          }
        },
        "required": [
          "product_id",
          "stock"
        ]
      },
      "UpdateProductVariantDto": {
        "type": "object",
        "properties": {
          "product_id": {
            "type": "number",
            "example": 1
          },
          "sku": {
            "type": "string",
            "example": "SKU-001"
          },
          "price_modifier": {
            "type": "number",
            "example": 0
          },
          "stock": {
            "type": "number",
            "example": 10
          },
          "is_active": {
            "type": "boolean",
            "example": true
          }
        }
      },
      "CreateAttributeTypeDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "example": "Color"
          },
          "slug": {
            "type": "string",
            "example": "color"
          },
          "applies_to": {
            "type": "string",
            "example": "products"
          }
        },
        "required": [
          "name",
          "slug"
        ]
      },
      "UpdateAttributeTypeDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "example": "Color"
          },
          "slug": {
            "type": "string",
            "example": "color"
          },
          "applies_to": {
            "type": "string",
            "example": "products"
          }
        }
      }
    }
  }
}